# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Our Empire - Satta Matka Application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/results:
    post:
      operationId: adminAddResult
      tags: [admin-results]
      summary: Add result for a market
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddResultRequest"
      responses:
        "201":
          description: Result added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
    get:
      operationId: adminGetResults
      tags: [admin-results]
      summary: Get all results
      security:
        - bearerAuth: []
      parameters:
        - name: marketId
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/Result"
                required: [results]
  /admin/results/{id}:
    put:
      operationId: adminUpdateResult
      tags: [admin-results]
      summary: Update result
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddResultRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
    delete:
      operationId: adminDeleteResult
      tags: [admin-results]
      summary: Delete result
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ADMIN TRANSACTIONS
  /admin/transactions/deposits:
    get:
      operationId: adminGetDeposits
      tags: [admin-transactions]
      summary: Get all deposit requests
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Deposits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTransactions"
  /admin/transactions/withdrawals:
    get:
      operationId: adminGetWithdrawals
      tags: [admin-transactions]
      summary: Get all withdrawal requests
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTransactions"
  /admin/transactions/deposits/{id}/approve:
    put:
      operationId: adminApproveDeposit
      tags: [admin-transactions]
      summary: Approve deposit
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/transactions/deposits/{id}/reject:
    put:
      operationId: adminRejectDeposit
      tags: [admin-transactions]
      summary: Reject deposit
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required: [reason]
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/transactions/withdrawals/{id}/complete:
    put:
      operationId: adminCompleteWithdrawal
      tags: [admin-transactions]
      summary: Complete withdrawal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/transactions/withdrawals/{id}/reject:
    put:
      operationId: adminRejectWithdrawal
      tags: [admin-transactions]
      summary: Reject withdrawal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required: [reason]
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ADMIN MARKETS
...
  /admin/settings:
    get:
      operationId: adminGetSettings
      tags: [admin-settings]
      summary: Get app settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
    put:
      operationId: adminUpdateSettings
      tags: [admin-settings]
      summary: Update app settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppSettings"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/settings/upi:
    post:
      operationId: adminAddUpi
      tags: [admin-settings]
      summary: Add UPI account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpiAccountRequest"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiAccount"
    get:
      operationId: adminGetUpiAccounts
      tags: [admin-settings]
      summary: Get all UPI accounts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: UPI accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  upiAccounts:
                    type: array
                    items:
                      $ref: "#/components/schemas/UpiAccount"
                required: [upiAccounts]
  /admin/settings/upi/{id}:
    put:
      operationId: adminUpdateUpi
      tags: [admin-settings]
      summary: Update UPI account
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpiAccountRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: adminDeleteUpi
      tags: [admin-settings]
      summary: Delete UPI account
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      required: [message, success]
...
    Result:
      type: object
      properties:
        id:
          type: string
        marketId:
          type: string
        marketName:
          type: string
        date:
          type: string
        openNumber:
          type: string
        closeNumber:
          type: string
        jodiNumber:
          type: string
        status:
          type: string
          enum: [scheduled, live, completed]
        createdAt:
          type: string
      required: [id, marketId, date, status]
    AddResultRequest:
      type: object
      properties:
        marketId:
          type: string
        date:
          type: string
        openNumber:
          type: string
        closeNumber:
          type: string
        jodiNumber:
          type: string
      required: [marketId, date]
    Transaction:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        type:
          type: string
          enum: [deposit, withdrawal, bet, win, referral, commission, adjustment]
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, rejected, cancelled]
        method:
          type: string
        utrNumber:
          type: string
        screenshotUrl:
          type: string
        adminNote:
          type: string
        bankDetails:
          type: object
          properties:
            accountNumber:
              type: string
            ifscCode:
              type: string
            accountHolderName:
              type: string
            upiId:
              type: string
        createdAt:
          type: string
        processedAt:
          type: string
      required: [id, type, amount, status, createdAt]
...
    UpiAccount:
      type: object
      properties:
        id:
          type: string
        upiId:
          type: string
        qrCodeUrl:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, upiId, isActive]
    UpiAccountRequest:
      type: object
      properties:
        upiId:
          type: string
        qrCodeUrl:
          type: string
        isActive:
          type: boolean
      required: [upiId]
...
    AppSettings:
      type: object
      properties:
        appName:
          type: string
        whatsappNumber:
          type: string
        telegramLink:
          type: string
        minDeposit:
          type: number
        minWithdrawal:
          type: number
        referralCommission:
          type: number
        maintenanceMode:
          type: boolean
        bettingLockMinutes:
          type: integer
      required: [appName]
...
    PaginatedTransactions:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [transactions, total, page, totalPages]
...